refactor: move api flickr fetching logic to flickr api client
