+++ backend/src/clients/__init__.py
@@ -1,0 +1,4 @@
+
+
+from .faceplusplus_api_client import FacePlusPlusApiClient
+from .flickr_api_client import FlickrApiClient
