+++ backend/src/fetch_photos.py
@@ -1,12 +1,12 @@
 from .services import service_locator
 from pprint import pprint
-flickr_service = service_locator.flickr_service
+photo_service = service_locator.photo_service
 
 page = 1
 
-response = flickr_service.fetch_sync(page)
+response = photo_service.fetch_sync(page)
 pprint(response)
-flickr_service._save_photos(response['photoset']['photo'])
+photo_service._save_photos(response['photoset']['photo'])
 
 for i in range((page+1), response.photoset.pages):
-    flickr_service.fetch_sync(page)
+    photo_service.fetch_sync(page)
