+++ backend/src/services/service_locator.py
@@ -3,11 +3,11 @@
 from __future__ import absolute_import, division, print_function
 
 from .faceplusplus_service import FacePlusPlusService
-from .flickr_service import FlickApiService
+from .photo_service import PhotoService
 
 
 class ServiceLocator:
 
     def __init__(self):
         self.faceplusplus_service = FacePlusPlusService()
-        self.flickr_service = FlickApiService()
+        self.photo_service = PhotoService()
